==> week1/05.pvalues.Rmd <==
---
title: "RNA-seq course- week1"
author: "Serhiy Naumenko"
date: "`r Sys.Date()`"
output:
  pdf_document:
    toc: yes
    keep_tex: yes
    latex_engine: xelatex
    number_sections: true
  html_document:
    code_folding: hide
    df_print: paged
    highlights: pygments
    number_sections: yes
    self_contained: yes
    theme: default
    toc: yes
    toc_float:
      collapsed: yes
      smooth_scroll: yes
mainfont: Tempora
fontsize: 10pt
urlcolor: blue
---

# Overview
- P-values and multiple testing correction
- https://bookdown.org/jgscott/DSGI/p-values.html
- https://uk.wikipedia.org/wiki/Метод_Монте-Карло
- https://uk.wikipedia.org/wiki/Біноміальний_розподіл



```{r include = FALSE}
library(tidyverse)
library(mosaic)
library(knitr)

set.seed(1)

ggplot2::theme_set(theme_light(base_size = 14))

opts_chunk[["set"]](
    cache = FALSE,
    dev = c("png", "pdf"),
    error = TRUE,
    highlight = TRUE,
    message = FALSE,
    prompt = FALSE,
    tidy = FALSE,
    warning = FALSE)

# https://www.rdocumentation.org/packages/mosaic/versions/1.8.4.2
# https://bookdown.org/jgscott/DSGI/p-values.html

```

\newpage

# Біноміальний розподіл
```{r}
# heads/tails аверс-реверс
p <- 0.5
n <- 25
bootstrap_n <- 10000

binomial_sim <- mosaic::do(bootstrap_n) * nflip(p = p, n = n)

ggplot(binomial_sim) + 
    geom_histogram(aes(x=nflip), binwidth = 1, boundary = 5) +
    geom_vline(xintercept = mean(binomial_sim$nflip), linetype = "dashed", color = "red") +
    geom_vline(xintercept = mean(binomial_sim$nflip) + sd(binomial_sim$nflip), linetype = "dashed", color = "yellow") +
    geom_vline(xintercept = mean(binomial_sim$nflip) - sd(binomial_sim$nflip), linetype = "dashed", color = "yellow") +
    scale_x_continuous(breaks = c(5, 7.5, 10, 12.5, 15, 17.5, 20))

```

* Mean= `r mean(binomial_sim$nflip)`
* Var = `r var(binomial_sim$nflip)`
* SD = `r sd(binomial_sim$nflip)`

\newpage

# Спостереження 
```{r}
# більш ніж 19
rare_event <- 19
sum(binomial_sim >= rare_event) / bootstrap_n
```
```{r}
binomial_sim$nflip %>% table()
```
\newpage

# P-значення
```{r}
ggplot(binomial_sim) + 
    geom_histogram(aes(x = nflip, fill = nflip >= 19), binwidth = 1, boundary = 5, closed = "left") +
    geom_vline(xintercept = mean(binomial_sim$nflip), linetype = "dashed", color = "red") +
    geom_vline(xintercept = mean(binomial_sim$nflip) + sd(binomial_sim$nflip), linetype = "dashed", color = "yellow") +
    geom_vline(xintercept = mean(binomial_sim$nflip) - sd(binomial_sim$nflip), linetype = "dashed", color = "yellow") +
    geom_vline(xintercept = 19, linetype = "solid", color = "black") +
    scale_x_continuous(breaks = c(4, 5, 7.5, 10, 12.5, 15, 17.5, 19, 20), 
                       limits = c(min(binomial_sim$nflip), max(binomial_sim$nflip))) + theme_bw()
#scale_y_log10()
```

- H0: монетка випадкова (p = 0.5) (нема біологічного сигнала)
- тестова статистика: скільки успіхів (heads) із 25 експериментів
- порахували розподіл ймовірності за методом Монте-Карло на 10,000 повторів
- оцінили, наскільки ймовірно, що H0 пояснює спостереження (більше 19) 

\newpage

# nflip >= 15
```{r}
ggplot(binomial_sim) + 
    geom_histogram(aes(x = nflip, fill = nflip >= 15), binwidth = 1, boundary = 5, closed = "left") +
    geom_vline(xintercept = mean(binomial_sim$nflip), linetype = "dashed", color = "red") +
    geom_vline(xintercept = mean(binomial_sim$nflip) + sd(binomial_sim$nflip), linetype = "dashed", color = "yellow") +
    geom_vline(xintercept = mean(binomial_sim$nflip) - sd(binomial_sim$nflip), linetype = "dashed", color = "yellow") +
    geom_vline(xintercept = 15, linetype = "solid", color = "black") +
    scale_x_continuous(breaks = c(4, 5, 7.5, 10, 12.5, 15, 17.5, 19, 20), 
                       limits = c(min(binomial_sim$nflip), max(binomial_sim$nflip))) + theme_bw()

```

```{r}
rare_event <- 15
sum(binomial_sim >= rare_event) / bootstrap_n
```

\newpage
# Кластер хвороби

- < 10 км від атомної станції: 5.8 випадків на 10,000: 47 / 80,515
- $>$ 30 км від атомної станції: 4.7 випадків на 10,000: 851 / 1,819,636
- incidence ratio: 5.8/4.7 = 1.23

- H0: IR = 4.7
- тестова статистика: кількість захворювань %
- розподіл тестової статистики за припущення H0

```{r}
sim_cancer <- do(10000)*nflip(n = 80515, prob = 0.00047)
ggplot(sim_cancer) + 
  geom_histogram(aes(x = nflip), binwidth = 1)
```

* Mean= `r mean(sim_cancer$nflip)`
* Var = `r var(sim_cancer$nflip)`
* SD = `r sd(sim_cancer$nflip)`
* P = `r sum(sim_cancer >= 47)/10000`

\newpage

```{r}
ggplot(sim_cancer) + 
    geom_histogram(aes(x = nflip, fill = nflip >= 47), binwidth = 1, boundary = 5, closed = "left") +
    geom_vline(xintercept = mean(sim_cancer$nflip), linetype = "dashed", color = "red") +
    geom_vline(xintercept = mean(sim_cancer$nflip) + sd(sim_cancer$nflip), linetype = "dashed", color = "yellow") +
    geom_vline(xintercept = mean(sim_cancer$nflip) - sd(sim_cancer$nflip), linetype = "dashed", color = "yellow") +
    geom_vline(xintercept = 47, linetype = "solid", color = "black") +
    scale_x_continuous(breaks = c(25, 32, 38, 44, 50), 
                       limits = c(min(sim_cancer$nflip), max(sim_cancer$nflip))) + theme_bw()

```

\newpage

# sessionInfo()
```{r}
sessionInfo()
```
